requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.1
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.4.4
beautifulsoup4==4.9.3
bootstrap4==0.1.0
boto==2.49.0
boto3==1.14.23
botocore==1.17.23
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==2.1.10
django-allauth==0.39.1
django-bootstrap-form==3.4
django-bootstrap4==2.3.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-embed-video==1.3.3
django-filer==1.7.1
django-filter==2.3.0
django-forms-bootstrap==3.1.0
django-jet==1.0.8
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==2.1.2
django-ratings==0.3.7
django-social-share==2.0.0
django-solo==1.1.3
django-storages==1.9.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
docutils==0.15.2
easy-thumbnails==2.7
filelock==3.0.12
gunicorn==19.9.0
html5lib==1.1
idna==2.8
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.3.0
oauthlib==3.0.1
packaging==20.3
path==13.1.0
path.py==12.4.0
pdfkit==0.6.1
pep8==1.7.1
Pillow==7.1.2
pluggy==0.13.1
Pmw==2.0.2
preppy==3.0.2
psycopg2-binary==2.8.4
py==1.8.1
pycodestyle==2.5.0
pylint==2.5.2
pyparsing==2.4.7
pyPdf==1.13
PyPDF2==1.26.0
pyRXP==2.2.1
pytest==5.4.2
python-dateutil==2.8.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2020.1
reportlab==3.5.44
requests==2.21.0
requests-oauthlib==1.2.0
rlextra==3.5.44
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
stripe==2.27.0
supervisor==4.1.0
toml==0.10.1
Unidecode==1.1.1
urllib3==1.24.2
virtualenv==20.0.25
wcwidth==0.1.9
webencodings==0.5.1
whitenoise==3.3.1
wrapt==1.12.1
xhtml2pdf==0.2.4
